Move some code to a .templates.h file.
